+++ src/anyio/_backends/_trio.py
@@ -674,6 +674,8 @@ class ConnectedUNIXDatagramSocket(
 
 
 class Event(BaseEvent):
+    __slots__ = ("__original",)
+
     def __new__(cls) -> Event:
         return object.__new__(cls)
 
@@ -695,6 +697,8 @@ class Event(BaseEvent):
 
 
 class Lock(BaseLock):
+    __slots__ = "_fast_acquire", "__original"
+
     def __new__(cls, *, fast_acquire: bool = False) -> Lock:
         return object.__new__(cls)
 
@@ -751,6 +755,8 @@ class Lock(BaseLock):
 
 
 class Semaphore(BaseSemaphore):
+    __slots__ = ("__original",)
+
     def __new__(
         cls,
         initial_value: int,
@@ -805,6 +811,8 @@ class Semaphore(BaseSemaphore):
 
 
 class CapacityLimiter(BaseCapacityLimiter):
+    __slots__ = ("__original",)
+
     def __new__(
         cls,
         total_tokens: float | None = None,