+++ src/anyio/_core/_tasks.py
@@ -55,7 +55,7 @@ class CancelScope:
         current thread
     """
 
-    __slots__ = ()
+    __slots__ = ("__weakref__",)
 
     def __new__(
         cls, *, deadline: float = math.inf, shield: bool = False